Bluetooth: GAP: Transition to stable state based on device
This patch adds check in BluetoothBondStateMachine to transition to
stable state on reception of BOND_STATE_NONE only for device that
are in pairing state and not for all devices. Without this patch it
was observed that state was transitioning to stable state for all
devices resulting in discarding of BOND_STATE_NONE from device that
was actually in pairing mode resulting in UI stuck in pairing forever.

Change-Id: Ia860d26a5f036c13d649434246ba6dea7b9807f2
CRs-Fixed: 584102