Bluetooth: GAP: Update bondstatemachine to handle pairing scenarios
This patch updates BluetoothBondStateMachine to properly handle concurrency
scenarios of pairing while pairing is ongoing with some remote device.
Without this patch it was observed that DUT was getting stuck in pairing
when unpairing was done for some paired devices and pairing timed out with
device for which pairing was being initiated.

Change-Id: I07d2bc323aebbeada9c930eb36c9855da59dd677
CRs-Fixed: 586935